Graphics: Fix filenames, cleanup include directives
