+++ game/game.vcxproj
@@ -401,6 +401,7 @@ xcopy /y "$(SolutionDir)external\SDL2_ttf-2.0.12\VisualC\external\lib\$(Platform
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="src\game\Graphics\ParticleGraphics.cpp" />
     <ClCompile Include="src\game\Graphics\RenderPasses\WaterTilesRenderPass.cpp" />
     <ClCompile Include="src\game\Graphics\RenderPasses\BackgroundRenderPass.cpp" />
     <ClCompile Include="src\game\Graphics\RenderPasses\ForegroundRenderPass.cpp" />
@@ -495,6 +496,7 @@ xcopy /y "$(SolutionDir)external\SDL2_ttf-2.0.12\VisualC\external\lib\$(Platform
     <ClCompile Include="src\game\script_implementation.c" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="src\game\Graphics\ParticleGraphics.hpp" />
     <ClInclude Include="src\game\Graphics\RenderPasses\WaterTilesRenderPass.hpp" />
     <ClInclude Include="src\game\Graphics\RenderPasses\BackgroundRenderPass.hpp" />
     <ClInclude Include="src\game\Graphics\RenderPasses\ForegroundRenderPass.hpp" />
