+++ game/game.vcxproj.filters
@@ -347,6 +347,9 @@
     <ClCompile Include="src\game\Graphics\RenderPasses\WaterTilesRenderPass.cpp">
       <Filter>Game Sources\Graphics\RenderPasses</Filter>
     </ClCompile>
+    <ClCompile Include="src\game\Graphics\ParticleGraphics.cpp">
+      <Filter>Game Sources\Graphics</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\game\egoboo.h">
@@ -648,6 +651,9 @@
     </ClInclude>
     <ClInclude Include="src\game\Graphics\RenderPasses\WaterTilesRenderPass.hpp">
       <Filter>Game Header Files\Graphics\RenderPasses</Filter>
+    </ClInclude>
+    <ClInclude Include="src\game\Graphics\ParticleGraphics.hpp">
+      <Filter>Game Header Files\Graphics</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
