+++ game/src/game/Entities/Common.hpp
@@ -19,6 +19,7 @@
 #pragma once
 
 #include "game/mesh.h"
+#include "game/physics.h"
 
 /**
  * @brief
