+++ game/src/game/Entities/Particle.hpp
@@ -29,6 +29,7 @@
 #include "game/egoboo.h"
 #include "game/graphic_prt.h"
 #include "game/Entities/Common.hpp"
+#include "game/Graphics/ParticleGraphics.hpp"
 #include "egolib/Graphics/Animation2D.hpp"
 #include "game/Physics/Collidable.hpp"
 #include "game/Physics/ParticlePhysics.hpp"
