+++ game/src/game/Graphics/ObjectGraphics.cpp
@@ -1,5 +1,6 @@
 #include "ObjectGraphics.hpp"
 #include "game/Entities/_Include.hpp"
+#include "game/graphic.h"
 #include "game/game.h" //only for character_swipe()
 
 namespace Ego
