+++ game/src/game/egoboo.c
@@ -25,6 +25,8 @@
 #include "game/game.h"
 #include "game/Entities/_Include.hpp"
 #include "game/Module/Module.hpp"
+#include "game/Graphics/CameraSystem.hpp"
+#include "game/graphic.h"
 
 //--------------------------------------------------------------------------------------------
